Example question (#16)
* feat(postback.py): add example questions with quick replies

Adds example questions for different categories (law, housing, club,
activity, course, admin procedure) using quick replies.  Also adds a
hint in the welcome message to reset the bot if it becomes unresponsive.

The addition of example questions with quick replies enhances user
experience by providing guidance on the types of questions the bot can
answer. The hint in the welcome message provides a solution for users
experiencing issues with the bot's responsiveness.

* fix(common.py): increase loading animation duration to 60 seconds

The default loading animation duration is increased to prevent premature
termination of the animation, especially during longer operations.

* feat: add example question to clubs

* Add law example question

* fix(dify.py): reduce max_retries to 1 to prevent long delays

The number of retries was reduced to prevent the bot from taking too long to respond when the Dify server is down.

* Add course example question